Move grpc logic to resolver
