+++ providers/openfeature-provider-flagd/src/openfeature/contrib/provider/flagd/provider.py
@@ -23,24 +23,13 @@
 
 import typing
 
-import grpc
-from google.protobuf.struct_pb2 import Struct
-
 from openfeature.evaluation_context import EvaluationContext
-from openfeature.exception import (
-    FlagNotFoundError,
-    GeneralError,
-    InvalidContextError,
-    ParseError,
-    TypeMismatchError,
-)
 from openfeature.flag_evaluation import FlagResolutionDetails
 from openfeature.provider.metadata import Metadata
 from openfeature.provider.provider import AbstractProvider
 
-from .config import Config
-from .flag_type import FlagType
-from .proto.schema.v1 import schema_pb2, schema_pb2_grpc
+from .config import Config, ResolverType
+from .resolvers.grpc import GrpcResolver
 
 T = typing.TypeVar("T")
 
@@ -54,6 +43,7 @@ class FlagdProvider(AbstractProvider):
         port: typing.Optional[int] = None,
         tls: typing.Optional[bool] = None,
         timeout: typing.Optional[int] = None,
+        resolver_type: typing.Optional[ResolverType] = None,
     ):
         """
         Create an instance of the FlagdProvider
@@ -68,14 +58,17 @@ class FlagdProvider(AbstractProvider):
             port=port,
             tls=tls,
             timeout=timeout,
+            resolver_type=resolver_type,
         )
 
-        channel_factory = grpc.secure_channel if tls else grpc.insecure_channel
-        self.channel = channel_factory(f"{self.config.host}:{self.config.port}")
-        self.stub = schema_pb2_grpc.ServiceStub(self.channel)
+        if self.config.resolver_type == ResolverType.GRPC:
+            self.resolver = GrpcResolver(self.config)
+        else:
+            raise ValueError("`resolver_type` parameter invalid")
 
     def shutdown(self) -> None:
-        self.channel.close()
+        if self.resolver:
+            self.resolver.shutdown()
 
     def get_metadata(self) -> Metadata:
         """Returns provider metadata"""
@@ -87,7 +80,9 @@ class FlagdProvider(AbstractProvider):
         default_value: bool,
         evaluation_context: typing.Optional[EvaluationContext] = None,
     ) -> FlagResolutionDetails[bool]:
-        return self._resolve(key, FlagType.BOOLEAN, default_value, evaluation_context)
+        return self.resolver.resolve_boolean_details(
+            key, default_value, evaluation_context
+        )
 
     def resolve_string_details(
         self,
@@ -95,7 +90,9 @@ class FlagdProvider(AbstractProvider):
         default_value: str,
         evaluation_context: typing.Optional[EvaluationContext] = None,
     ) -> FlagResolutionDetails[str]:
-        return self._resolve(key, FlagType.STRING, default_value, evaluation_context)
+        return self.resolver.resolve_string_details(
+            key, default_value, evaluation_context
+        )
 
     def resolve_float_details(
         self,
@@ -103,7 +100,9 @@ class FlagdProvider(AbstractProvider):
         default_value: float,
         evaluation_context: typing.Optional[EvaluationContext] = None,
     ) -> FlagResolutionDetails[float]:
-        return self._resolve(key, FlagType.FLOAT, default_value, evaluation_context)
+        return self.resolver.resolve_float_details(
+            key, default_value, evaluation_context
+        )
 
     def resolve_integer_details(
         self,
@@ -111,7 +110,9 @@ class FlagdProvider(AbstractProvider):
         default_value: int,
         evaluation_context: typing.Optional[EvaluationContext] = None,
     ) -> FlagResolutionDetails[int]:
-        return self._resolve(key, FlagType.INTEGER, default_value, evaluation_context)
+        return self.resolver.resolve_integer_details(
+            key, default_value, evaluation_context
+        )
 
     def resolve_object_details(
         self,
@@ -119,76 +120,6 @@ class FlagdProvider(AbstractProvider):
         default_value: typing.Union[dict, list],
         evaluation_context: typing.Optional[EvaluationContext] = None,
     ) -> FlagResolutionDetails[typing.Union[dict, list]]:
-        return self._resolve(key, FlagType.OBJECT, default_value, evaluation_context)
-
-    def _resolve(
-        self,
-        flag_key: str,
-        flag_type: FlagType,
-        default_value: T,
-        evaluation_context: typing.Optional[EvaluationContext],
-    ) -> FlagResolutionDetails[T]:
-        context = self._convert_context(evaluation_context)
-        call_args = {"timeout": self.config.timeout}
-        try:
-            if flag_type == FlagType.BOOLEAN:
-                request = schema_pb2.ResolveBooleanRequest(  # type:ignore[attr-defined]
-                    flag_key=flag_key, context=context
-                )
-                response = self.stub.ResolveBoolean(request, **call_args)
-            elif flag_type == FlagType.STRING:
-                request = schema_pb2.ResolveStringRequest(  # type:ignore[attr-defined]
-                    flag_key=flag_key, context=context
-                )
-                response = self.stub.ResolveString(request, **call_args)
-            elif flag_type == FlagType.OBJECT:
-                request = schema_pb2.ResolveObjectRequest(  # type:ignore[attr-defined]
-                    flag_key=flag_key, context=context
-                )
-                response = self.stub.ResolveObject(request, **call_args)
-            elif flag_type == FlagType.FLOAT:
-                request = schema_pb2.ResolveFloatRequest(  # type:ignore[attr-defined]
-                    flag_key=flag_key, context=context
-                )
-                response = self.stub.ResolveFloat(request, **call_args)
-            elif flag_type == FlagType.INTEGER:
-                request = schema_pb2.ResolveIntRequest(  # type:ignore[attr-defined]
-                    flag_key=flag_key, context=context
-                )
-                response = self.stub.ResolveInt(request, **call_args)
-            else:
-                raise ValueError(f"Unknown flag type: {flag_type}")
-
-        except grpc.RpcError as e:
-            code = e.code()
-            message = f"received grpc status code {code}"
-
-            if code == grpc.StatusCode.NOT_FOUND:
-                raise FlagNotFoundError(message) from e
-            elif code == grpc.StatusCode.INVALID_ARGUMENT:
-                raise TypeMismatchError(message) from e
-            elif code == grpc.StatusCode.DATA_LOSS:
-                raise ParseError(message) from e
-            raise GeneralError(message) from e
-
-        # Got a valid flag and valid type. Return it.
-        return FlagResolutionDetails(
-            value=response.value,
-            reason=response.reason,
-            variant=response.variant,
+        return self.resolver.resolve_object_details(
+            key, default_value, evaluation_context
         )
-
-    def _convert_context(
-        self, evaluation_context: typing.Optional[EvaluationContext]
-    ) -> Struct:
-        s = Struct()
-        if evaluation_context:
-            try:
-                s["targetingKey"] = evaluation_context.targeting_key
-                s.update(evaluation_context.attributes)
-            except ValueError as exc:
-                message = (
-                    "could not serialize evaluation context to google.protobuf.Struct"
-                )
-                raise InvalidContextError(message) from exc
-        return s
